Enable IPv6
As now several VPN provider assign a IPv6 address, their config files would result in a failure when wireguard tries to bind the address to the wg0 interface as v6 is dissabled by default. As a result the wireguard container starts without a active VPN connection what could lead to a unnoticed ip leak.